refactor: RacingView 객체로 책임 이동
