+++ src/main/java/racingcar/RacingManager.java
@@ -9,12 +9,14 @@ public class RacingManager {
     private final Cars cars;
     private final int round;
     private final List<List<CarDto>> records;
+    private final RacingViewManager racingViewManager;
 
     public RacingManager(Cars cars, int round) {
         this.cars = cars;
         this.round = round;
         records = new ArrayList<>();
         records.add(cars.getRecord());
+        racingViewManager = new RacingViewManager(cars, records);
     }
 
     public static RacingManager of(List<String> carNames, int round) {
@@ -27,29 +29,6 @@ public class RacingManager {
             records.add(cars.getRecord());
         }
 
-        printResult();
-    }
-
-    public void printResult() {
-        System.out.println("실행 결과");
-        printRecords();
-        printWinners();
-    }
-
-    private void printRecords() {
-        for (List<CarDto> record : records) {
-            printCarRecord(record);
-            System.out.println();
-        }
-    }
-
-    private static void printCarRecord(List<CarDto> record) {
-        for (CarDto carRecord : record) {
-            System.out.println(carRecord.getName() + " : " + "-".repeat(carRecord.getDistance() + 1));
-        }
-    }
-
-    private void printWinners() {
-        System.out.println(String.join(",", cars.findWinnerNames()) + WINNERS_STRING);
+        racingViewManager.printResult();
     }
 }
